add multitenant support
If the `multi_tenant` flag is set in the `.proto`, a `ID` field will be
added to the gorm model, and the CRUD methods will be modified to
include tenancy checks.

*Limitations:* if an `id` field isn't present on the message, then the
`DefaultUpdateX` function can't be created. This could be extended to
have better support, but for this iteration the `id` convention will be
required.