feat: add partial support for composite primary keys
Currently, Only single PK was getting handled. This PR attmepts to
add the support for composite primary keys, so that generated methods
are predictable.

Currently, only read, list & delete is handled in the current PR.
Additional method support will be added in the upcoming PRs.

Added department user proto to demostrate the same.